Adding ARM support
Currently, any attempt to build this on an ARM (aarch64) machine will fail because of an "asm volatile" memory fence that has an x86 specific integer register constraint.  I admittedly don't fully understand why this is necessary, so I #ifdef'd it out on ARM to use a general register constraint.  This seems to work nicely, and should be directly equivalent:

https://developer.arm.com/documentation/dui0774/l/armclang-Inline-Assembler/Inline-assembly-constraint-strings/Constraint-codes-for-AArch64-state?lang=en
https://gcc.gnu.org/onlinedocs/gcc-14.3.0/gcc/Simple-Constraints.html
https://gcc.gnu.org/onlinedocs/gcc-14.3.0/gcc/Machine-Constraints.html